add support for mangafox mirror
The owners of mangafox have seemingly decided in their infinite wisdom to create ANOTHER site just an hr or so ago. This time it's called `fanfox.net`.

Aside from annoying the crap out of me due to them again reuploading every damn manga and spamming the rss feed with 500 something notifications within the hr, I suspect there's a chance that this might be in prep for another domain move, which if so, this script change will enable both domains to be supported, for a smoother transition for users. 

I sincerely hope they stop moving domains.